requirement.txt
absl-py==0.7.1
astor==0.7.1
backports.functools-lru-cache==1.5
backports.weakref==1.0.post1
bleach==1.5.0
cloudpickle==0.8.0
cycler==0.10.0
dask==1.1.4
decorator==4.3.2
enum34==1.1.6
funcsigs==1.0.2
futures==3.2.0
gast==0.2.2
grpcio==1.19.0
h5py==2.9.0
html5lib==0.9999999
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
Markdown==3.1
matplotlib==2.2.4
mock==2.0.0
networkx==2.2
numpy==1.16.2
opencv-python==4.0.0.21
pbr==5.1.3
Pillow==5.4.1
protobuf==3.7.0
pydot==1.4.1
pyparsing==2.3.1
Pyste==0.9.10
python-dateutil==2.8.0
pytz==2018.9
PyWavelets==1.0.2
scikit-image==0.14.2
scipy==1.2.1
six==1.12.0
subprocess32==3.5.3
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.4.0
tensorflow-tensorboard==0.4.0
termcolor==1.1.0
toolz==0.9.0
Werkzeug==0.15.1
